add td error logging for envelope algos, change lava gap test envs configurations, add a-star algorithm to find best path

+++ envs/mo_lava_gap/mo_lava_gap.py
@@ -268,7 +268,7 @@ class MOLavaGapDR(MiniGridEnv, DREnv):
         ])
 
         full_grid = full_grid.flatten()
-        return full_grid/ 1.
+        return full_grid.astype(np.float32) / 1.
     
     def rgb_observation(self):
         env = self.unwrapped
